init steps to separating search and execution
